// src/app/core/api/generated/model/product-request.ts
/**
 * Seggi App API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProductRequest { 
    name: string;
    price: number;
    weight?: number;
    categoryId?: number;
}