// src/app/core/api/generated/model/product-response.ts
/**
 * Seggi App API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProductResponse { 
    idProduct?: number;
    name?: string;
    price?: number;
    weight?: number;
    categoryId?: number;
    categoryName?: string;
}